Moved traits CausableGraph and CausableGraphReasoning into two seperate
